remove unnecessary dependencies
