+++ pom.xml
@@ -110,13 +110,6 @@
             <version>1.18</version>
         </dependency>
 
-        <!-- https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-rio-api -->
-        <dependency>
-            <groupId>org.eclipse.rdf4j</groupId>
-            <artifactId>rdf4j-rio-api</artifactId>
-            <version>${rdf4j.version}</version>
-        </dependency>
-
         <!-- https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-rio-ntriples -->
         <dependency>
             <groupId>org.eclipse.rdf4j</groupId>
@@ -124,25 +117,6 @@
             <version>${rdf4j.version}</version>
         </dependency>
 
-        <!-- https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-rio-ntriples -->
-        <dependency>
-            <groupId>org.eclipse.rdf4j</groupId>
-            <artifactId>rdf4j-rio-nquads</artifactId>
-            <version>${rdf4j.version}</version>
-        </dependency>
-
-        <!-- https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-rio-ntriples -->
-        <dependency>
-            <groupId>org.eclipse.rdf4j</groupId>
-            <artifactId>rdf4j-rio-rdfxml</artifactId>
-            <version>${rdf4j.version}</version>
-        </dependency>
-        <!-- https://mvnrepository.com/artifact/org.eclipse.rdf4j/rdf4j-rio-ntriples -->
-        <dependency>
-            <groupId>org.eclipse.rdf4j</groupId>
-            <artifactId>rdf4j-rio-turtle</artifactId>
-            <version>${rdf4j.version}</version>
-        </dependency>
 
         <!-- used for time measurements-->
         <dependency>
@@ -170,16 +144,6 @@
             <version>2.0</version>
         </dependency>
         <dependency>
-            <groupId>org.scalactic</groupId>
-            <artifactId>scalactic_${scala.compat.version}</artifactId>
-            <version>3.0.5</version>
-        </dependency>
-        <dependency> <!-- required for scalactic -->
-            <groupId>org.scala-lang.modules</groupId>
-            <artifactId>scala-xml_${scala.compat.version}</artifactId>
-            <version>1.1.1</version>
-        </dependency>
-        <dependency>
             <groupId>com.lihaoyi</groupId>
             <artifactId>fastparse_${scala.compat.version}</artifactId>
             <version>2.0.4</version>
@@ -191,11 +155,6 @@
             <version>1.3.5</version>
         </dependency>
 
-        <dependency>
-            <groupId>org.tukaani</groupId>
-            <artifactId>xz</artifactId>
-            <version>1.8</version>
-        </dependency>
         <!-- https://mvnrepository.com/artifact/com.github.luben/zstd-jni notwendig fuer zstd compression-->
         <dependency>
             <groupId>com.github.luben</groupId>
@@ -203,18 +162,20 @@
             <version>1.4.3-1</version>
         </dependency>
 
-
         <!-- test dependencies -->
+
+        <!-- need for integration tests -->
+        <dependency>
+            <groupId>org.apache.maven.shared</groupId>
+            <artifactId>maven-verifier</artifactId>
+            <version>1.7.2</version>
+            <scope>test</scope>
+        </dependency>
+        <!-- needed for unit tests -->
         <dependency>
             <groupId>org.apache.maven.plugin-testing</groupId>
             <artifactId>maven-plugin-testing-harness</artifactId>
             <version>3.3.0</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.maven</groupId>
-            <artifactId>maven-core</artifactId>
-            <version>${maven.plugin.tools.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -224,34 +185,9 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.apache.maven</groupId>
-            <artifactId>maven-model</artifactId>
-            <version>${maven.plugin.tools.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.maven</groupId>
-            <artifactId>maven-aether-provider</artifactId>
-            <version>3.2.5</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.codehaus.plexus</groupId>
-            <artifactId>plexus-utils</artifactId>
-            <version>3.0.15</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
             <groupId>org.scalatest</groupId>
             <artifactId>scalatest_${scala.compat.version}</artifactId>
             <version>3.0.5</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.maven.shared</groupId>
-            <artifactId>maven-verifier</artifactId>
-            <version>1.7.2</version>
             <scope>test</scope>
         </dependency>
         <!-- test dependencies -->
