+++ src/test/resources/sample-pom.xml
@@ -8,6 +8,7 @@
     <version>1.3-SNAPSHOT</version>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <metadata.ipfsSettings.isInDocker>true</metadata.ipfsSettings.isInDocker>
     </properties>
 
     <build>
